Crate jane_eyre

By Currer Bell

Modules

config
Configuration options for customizing the behavior of the provided panic and error reporting hooks

eyre
This library provides eyre::Report, a trait object based error handling type for easy idiomatic error handling and reporting in Rust applications.

section
Helpers for adding custom sections to error reports

Structs

Context
A custom handler type for eyre::Report which provides colorful error reports and tracing-error support.

Handler
A custom handler type for eyre::Report which provides colorful error reports and tracing-error support.

IndentedSection
An indenteted section with a header for an error report

Report
The core error reporting type of the library, a wrapper around a dynamic error reporting type.

Traits

Help
A helper trait for attaching informational sections to error reports to be displayed after the chain of errors

Section
A helper trait for attaching informational sections to error reports to be displayed after the chain of errors

SectionExt
Extension trait for constructing sections with commonly used formats

Functions

install
Install the default panic and error report hooks

Type Definitions

Result
type alias for Result<T, Report>